// src/app/interfaces/studentsInterface.ts
export interface Students {
    firstName:      string;
    lastName:       string;
    fatherName:     string;
    dateOfBirth:    Date;
    subjects:       string[];
    phoneNumber:    string;
    address:        string;
    gender:         string;
    email:          string;
    userName:       string;
    password:       string;
    activitystatus: string;
    name:           string;
    id?:             number;
}